requirements.txt
streamlit
deepface
gspread
google-auth
numpy
Pillow
scikit-learn
